refactor: restructure core services and monitors architecture
- Move monitors from separate modules to module/core/monitors/
  - download_monitor.py -> core/monitors/download_monitor.py
  - notification_monitor.py -> core/monitors/notification_monitor.py
  - rename_monitor.py -> core/monitors/rename_monitor.py
- Split large services.py into modular service files in core/services/
  - BaseService -> base_services.py
  - DownloadService -> download_service.py
  - RSSService -> rss_service.py
  - RenamerService -> renamer_service.py (moved from core/)
- Update imports and references across affected modules
- Remove obsolete notification config and manager modules
- Clean up unused status reporting methods
