+++ backend/src/module/core/monitors/__init__.py
@@ -1,0 +1,10 @@
+from .download_monitor import DownloadMonitor,download_monitor
+from .notification_monitor import NotificationMonitor
+from .rename_monitor import RenameMonitor
+
+__all__ = [
+    "DownloadMonitor",
+    "download_monitor",
+    "NotificationMonitor",
+    "RenameMonitor",
+]
