+++ backend/src/module/core/services/__init__.py
@@ -1,0 +1,6 @@
+from .base_services import BaseService
+from .download_service import DownloadService
+from .renamer_service import RenamerService
+from .rss_service import RSSService
+
+__all__ = ["BaseService", "DownloadService", "RenamerService", "RSSService"]
