+++ backend/src/module/downloader/__init__.py
@@ -1,5 +1,4 @@
 from .download_client import Client, DownloadClient
-from .download_monitor import DownloadMonitor, download_monitor
 from .download_queue import DownloadController, DownloadQueue
 
 __all__ = [
@@ -7,6 +6,4 @@ __all__ = [
     "DownloadQueue",
     "DownloadController",
     "Client",
-    "DownloadMonitor",
-    "download_monitor",
 ]
