+++ backend/src/module/notification/__init__.py
@@ -1,18 +1,8 @@
-from .config import NotificationConfigManager, NotificationSettings
-from .manager import NotificationConfig, NotificationManager, NotificationType
-from .notification import NotificationProcessor, NotificationSender, PostNotification
-from .notification_monitor import NotificationMonitor
+from .notification import NotificationProcessor, PostNotification
 
 __all__ = [
     # 核心通知类
     "PostNotification",
     "NotificationProcessor",
-    "NotificationSender",
-    "NotificationMonitor",
     # 管理器和配置
-    "NotificationManager",
-    "NotificationConfig",
-    "NotificationConfigManager",
-    "NotificationSettings",
-    "NotificationType",
 ]
