+++ backend/src/module/notification/notification.py
@@ -1,22 +1,20 @@
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING
+from enum import Enum
 
 from module.conf import settings
 from module.database import Database
 from module.models import Notification
 from module.network import load_image
-from module.notification.manager import (
-    NotificationConfig,
-    NotificationManager,
-    NotificationType,
-)
+import importlib
 
-if TYPE_CHECKING:
-    pass
+from module.notification.plugin.log import Notifier
 
 logger = logging.getLogger(__name__)
+
+
+NOTIFICATON_TYPE = ["tetegram", "bark", "server_chan", "wecom", "log"]
 
 
 class NotificationProcessor:
@@ -51,7 +49,6 @@ class NotificationProcessor:
             message=notify.message,
         )
         # 想了想,对每个通知直接传file和post_link, 这样就可以直接用来发送了
-
         # 生成默认消息
         if processed.episode:
             processed.message = f"番剧名称：{processed.title}\n季度：第{processed.season}季\n更新集数：第{processed.episode}集"
@@ -69,66 +66,35 @@ class NotificationProcessor:
 
 class PostNotification:
     """
-    现代化的通知发送类
     使用新的 NotificationManager 进行管理
-    保持向后兼容的接口
     """
 
     def __init__(self) -> None:
         self.processor = NotificationProcessor()
-        self.manager = self._create_manager_from_settings()
         self.logger = logging.getLogger(self.__class__.__name__)
+        self.notifier = None
 
-    def _create_manager_from_settings(self) -> NotificationManager:
+    def initialize(self) -> None:
         """根据设置创建通知管理器"""
-        configs = []
+        self.notifier = self._get_notifier()
+        self.notifier.initialize()
 
-        if settings.notification.enable:
-            # 解析多个chat_id
-            chat_ids = [
-                cid.strip()
-                for cid in settings.notification.chat_id.split(",")
-                if cid.strip()
-            ]
-
-            # 确定通知类型
-            try:
-                notification_type = NotificationType(settings.notification.type)
-            except ValueError:
-                logger.warning(
-                    f"未知的通知类型: {settings.notification.type}，回退到log"
-                )
-                notification_type = NotificationType.LOG
-
-            # 为每个chat_id创建配置
-            if chat_ids and notification_type in [
-                NotificationType.TELEGRAM,
-                NotificationType.WECOM,
-            ]:
-                for chat_id in chat_ids:
-                    config = NotificationConfig(
-                        type=notification_type,
-                        token=settings.notification.token,
-                        chat_id=chat_id,
-                        enabled=True,
-                    )
-                    configs.append(config)
-            else:
-                # 单个配置
-                config = NotificationConfig(
-                    type=notification_type,
-                    token=settings.notification.token,
-                    enabled=True,
-                )
-                configs.append(config)
-        else:
-            # 禁用时使用日志通知
-            config = NotificationConfig(
-                type=NotificationType.LOG, token="", enabled=True
+    def _get_notifier(self):
+        notification_type = settings.notification.type
+        package_path = f"module.notification.plugin.{notification_type}"
+        try:
+            notification_module = importlib.import_module(package_path)
+        except ImportError as e:
+            logger.error(f"加载通知器失败: {notification_type} - {e}")
+            notification_module = importlib.import_module(
+                "module.notification.plugin.log"
             )
-            configs.append(config)
-
-        return NotificationManager(configs)
+            logger.warning(
+                f"使用默认日志通知器: {notification_module.Notifier.__name__}"
+            )
+        Notifier = notification_module.Notifier
+        logger.debug(f"加载通知器: {Notifier.__name__}")
+        return Notifier()
 
     async def send(self, notify: Notification) -> bool:
         """
@@ -144,18 +110,9 @@ class PostNotification:
         processed_notify = await self.processor.process_notification(notify)
 
         # 发送通知
-        results = await self.manager.send_notification(processed_notify)
+        result = await self.notifier.post_msg(processed_notify)
 
-        # 只要有一个成功就返回True（保持向后兼容）
-        return any(results.values()) if results else False
-
-    def get_status(self) -> dict:
-        """获取通知系统状态"""
-        return self.manager.get_status()
-
-
-# 为了向后兼容，保留旧的类名
-NotificationSender = PostNotification
+        return result
 
 
 if __name__ == "__main__":
@@ -169,11 +126,11 @@ if __name__ == "__main__":
     title = "败犬"
     link = "posters/aHR0cHM6Ly9pbWFnZS50bWRiLm9yZy90L3Avdzc4MC9wYWRSbWJrMkxkTGd1ZGg1Y0xZMG85VEZ6aEkuanBn"
     nt = Notification(title=title, season=1, episode="1", poster_path=link)
+    sender = PostNotification()
+    sender.initialize()
 
     async def test():
-        sender = PostNotification()
         success = await sender.send(nt)
         print(f"发送结果: {success}")
-        print(f"系统状态: {sender.get_status()}")
 
     asyncio.run(test())
