+++ backend/src/module/notification/plugin/bark.py
@@ -1,5 +1,8 @@
-from typing import Any, Dict
+from typing import Any
 
+from pydantic import BaseModel, Field
+
+from module.conf import get_plugin_config
 from module.models import Notification
 from module.network import RequestContent
 from module.utils.cache_image import str_to_url
@@ -7,46 +10,72 @@ from module.utils.cache_image import str_to_url
 from .base_notifier import BaseNotifier
 
 
+class Config(BaseModel):
+    token: str = Field(
+        default="",
+        alias="token",
+        description="Telegram bot token, can be obtained from BotFather",
+    )
+
 class Notifier(BaseNotifier):
     """Bark 通知器"""
 
-    def __init__(self, token: str, **kwargs):
-        super().__init__(token, **kwargs)
+    def __init__(self):
+        super().__init__()
         self.notification_url = "https://api.day.app/push"
 
-    def format_message(self, notify: Notification) -> Dict[str, Any]:
-        """格式化 Bark 通知消息"""
-        data = super().format_message(notify)
+    def initialize(self) -> None:
+        self.config: Config = get_plugin_config(Config(), "notification")
+        self.token: str = self.config.token.strip()
+        self.photo_url: str = (
+            f"https://api.telegram.org/bot{self.config.token}/sendPhoto"
+        )
+        self.message_url: str = (
+            f"https://api.telegram.org/bot{self.config.token}/sendMessage"
+        )
 
-        # 处理海报路径
+    def format_message(self, notify: Notification) -> dict[str, Any]:
+        """格式化 Bark 通知消息"""
         poster_path = notify.poster_path
         if poster_path:
             if "/" in poster_path:
                 poster_path = str_to_url(poster_path.split("/")[-1])
 
-        return {**data, "poster_path": poster_path}
+        notify.poster_path = poster_path
 
     async def post_msg(self, notify: Notification) -> bool:
         """发送 Bark 通知"""
         try:
-            message_data = self.format_message(notify)
+            self.format_message(notify)
 
             data = {
-                "title": message_data["title"],
-                "body": message_data["message"],
+                "title": notify.title,
+                "body": notify.message,
                 "device_key": self.token,
             }
 
             # 添加图标
-            if message_data["poster_path"]:
-                data["icon"] = message_data["poster_path"]
+            if notify.poster_path:
+                data["icon"] = notify.poster_path
 
             async with RequestContent() as req:
                 resp = await req.post_data(self.notification_url, data)
 
             self.logger.debug(f"Bark notification response: {resp.status_code}")
-            return resp and resp.status_code == 200
+            resp.raise_for_status()
+            return True
 
         except Exception as e:
             self.logger.error(f"Bark 通知发送失败: {e}")
             return False
+if __name__ == "__main__":
+    # 示例用法
+    #https://api.day.app/sG7GDikc5iEu5c5iXEuDa4/Icon?icon=https://day.app/assets/images/avatar.jpg
+    notifier = Notifier(token="sG7GDikc5iEu5c5iXEuDa4")
+    notification = Notification(
+        title="测试通知",
+        message="这是一个测试消息",
+        poster_path="/path/to/poster.jpg"
+    )
+    success = notifier.post_msg(notification)
+    print(f"通知发送成功: {success}")
