+++ backend/src/module/notification/plugin/base_notifier.py
@@ -12,9 +12,8 @@ class BaseNotifier(ABC):
     通知器基类，定义统一的通知接口
     """
 
-    def __init__(self, token: str, **kwargs):
-        self.token = token
-        self.config = kwargs
+    def __init__(self):
+        self.chat_id:str = ""
         self.logger = logging.getLogger(self.__class__.__name__)
 
     @abstractmethod
@@ -29,28 +28,6 @@ class BaseNotifier(ABC):
             bool: 发送是否成功
         """
         pass
-
-    def format_message(self, notify: Notification) -> Dict[str, Any]:
-        """
-        格式化通知消息，子类可重写此方法
-
-        Args:
-            notify: 通知对象
-
-        Returns:
-            Dict[str, Any]: 格式化后的消息数据
-        """
-        message = notify.message
-        if not message and notify.episode:
-            message = f"番剧名称：{notify.title}\n季度：第{notify.season}季\n更新集数：第{notify.episode}集"
-
-        return {
-            "title": notify.title,
-            "message": message,
-            "poster_path": notify.poster_path,
-            "season": notify.season,
-            "episode": notify.episode,
-        }
 
     async def send_with_retry(self, notify: Notification, max_retries: int = 3) -> bool:
         """
