+++ backend/src/module/notification/plugin/log.py
@@ -8,17 +8,12 @@ class Notifier(BaseNotifier):
 
     def __init__(self, token: str = "", **kwargs):
         # 日志通知器不需要token，但保持接口一致
-        super().__init__(token, **kwargs)
+        super().__init__()
 
     async def post_msg(self, notify: Notification) -> bool:
         """输出通知到日志"""
         try:
-            message_data = self.format_message(notify)
-            message = (
-                message_data["message"]
-                or f"{notify.title} S{notify.season}E{notify.episode}"
-            )
-
+            message = notify.message
             self.logger.info(f"[Notification] {message}")
             return True
